Rename resourceHandler to resourceHelper - Rename factory files & variables - Rename UVP resourceHelper file & variables - Rename variables in OgcResourceService
